cloud_function/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
requests
functions-framework
google-cloud-firestore